authoritiy controller 추가

+++ src/main/java/com/nhnacademy/taskApi/repository/ProjectRepository.java
@@ -3,5 +3,6 @@ package com.nhnacademy.taskApi.repository;
 import com.nhnacademy.taskApi.domain.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
+public interface ProjectRepository extends JpaRepository<Project, Long> {
+
 }
